requirements.txt
# --- Web / API ---
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.2
typing_extensions==4.12.2

# --- Segurança / Autenticação ---
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.1.0.post1

# --- MongoDB ---
motor==3.4.0
pymongo==4.6.1
dnspython==2.6.1

# --- Groq ---
groq==0.12.0
httpx==0.23.3

# --- Cohere (embeddings) ---
cohere==5.11.0
numpy==1.26.4

# --- Outras dependências auxiliares ---
requests==2.32.3
urllib3==2.2.3
sniffio==1.3.0
anyio==4.4.0
h11==0.14.0
